Account + Client + Company + User

+++ BE/freelancer/src/main/java/org/example/freelancer/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.example.freelancer.repository;
-import org.example.freelancer.model.User;
+import org.example.freelancer.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
